Update execution model to use Error interface

+++ web_app/src/lib/server/db/models/ExecutionModel.ts
@@ -1,20 +1,36 @@
-import { Model, type Execution } from '$lib/types';
+import { Model, type Execution, type Error, isExecutionWithoutId } from '$lib/types';
 import { type Collection, ObjectId } from 'mongodb';
 import db from '$db/database';
 
 export default class ExecutionModel extends Model<Execution> {
 	collection: Collection = db.collection('executions');
-	async create(execution: Execution): Promise<ObjectId | null> {
+	async create(execution: Execution): Promise<ObjectId | Error> {
 		try {
+			if (!isExecutionWithoutId(execution)) {
+				return {
+					code: 400,
+					message: 'Invalid website record without ID'
+				} as Error;
+			}
+
 			const result = await this.collection.insertOne(execution);
 			return result.insertedId;
 		} catch (error) {
-			console.log(error);
-			return null;
+			return {
+				code: 500,
+				message: 'Failed to create execution'
+			} as Error;
 		}
 	}
-	async getByID(id: string): Promise<Execution | null> {
+	async getByID(id: string): Promise<Execution | Error> {
 		try {
+			if (!ObjectId.isValid(id)) {
+				return {
+					code: 400,
+					message: 'Invalid Mongo ID'
+				} as Error;
+			}
+
 			const result = await this.collection.findOne({ _id: new ObjectId(id) });
 
 			if (result) {
@@ -32,38 +48,81 @@ export default class ExecutionModel extends Model<Execution> {
 					title: result.title
 				} as Execution;
 			}
-			return null;
+
+			return {
+				code: 404,
+				message: 'Execution not found'
+			} as Error;
 		} catch (error) {
-			console.log(error);
-            return null;
+			return {
+				code: 500,
+				message: 'Failed to get execution'
+			} as Error;
 		}
 	}
-	async update(id: string, updatedItem: Execution): Promise<Execution | null> {
+	async update(id: string, updatedItem: Execution): Promise<Execution | Error> {
 		try {
+			if (!ObjectId.isValid(id)) {
+				return {
+					code: 400,
+					message: 'Invalid Mongo ID'
+				} as Error;
+			}
+
+			if (!isExecutionWithoutId(updatedItem)) {
+				return {
+					code: 400,
+					message: 'Invalid execution without ID'
+				} as Error;
+			}
+
 			const result = await this.collection.updateOne(
 				{ _id: new ObjectId(id) },
 				{ $set: updatedItem }
 			);
 			if (result.modifiedCount > 0) {
-				return updatedItem;
+				return {
+					id: new ObjectId(id),
+					...updatedItem
+				} as Execution;
 			}
-			return null;
+			return {
+				code: 404,
+				message: 'Execution not found or no changes were made'
+			} as Error;
 		} catch (error) {
-			console.log(error);
-			return null;
+			return {
+				code: 500,
+				message: 'Failed to update execution'
+			} as Error;
 		}
 	}
-	async delete(id: string): Promise<boolean> {
+	async delete(id: string): Promise<boolean | Error> {
 		try {
+			if (!ObjectId.isValid(id)) {
+				return {
+					code: 400,
+					message: 'Invalid Mongo ID'
+				} as Error;
+			}
+
 			const result = await this.collection.deleteOne({ _id: new ObjectId(id) });
-			return result.deletedCount > 0;
+			if (result.deletedCount > 0) {
+				return true;
+			}
+			return {
+				code: 404,
+				message: 'Execution not found'
+			} as Error;
 		} catch (error) {
-			console.log(error);
-			return false;
+			return {
+				code: 500,
+				message: 'Failed to delete website record'
+			} as Error;
 		}
 	}
 
-	async getAll(): Promise<Execution[]> {
+	async getAll(): Promise<Execution[] | Error> {
 		try {
 			const result = await this.collection.find({ root: true }).toArray();
 			return result.map(
@@ -83,15 +142,33 @@ export default class ExecutionModel extends Model<Execution> {
 					} as Execution)
 			);
 		} catch (error) {
-			console.log(error);
-			return [];
+			return {
+				code: 500,
+				message: 'Failed to get all root executions'
+			} as Error;
 		}
 	}
 
-    async getAllByOwnerIdAndGroupId(ownerId: string, groupId: number): Promise<Execution[]> {
-        try {
-            const result = await this.collection.find({ ownerId: new ObjectId(ownerId), groupId }).toArray();
-            return result.map(
+	async getAllByOwnerIdAndGroupId(ownerId: string, groupId: number): Promise<Execution[] | Error> {
+		try {
+			if (!ObjectId.isValid(ownerId)) {
+				return {
+					code: 400,
+					message: 'Invalid Mongo ID'
+				} as Error;
+			}
+
+			if (typeof groupId !== 'number' || groupId < 0) {
+				return {
+					code: 400,
+					message: 'Invalid group ID'
+				} as Error;
+			}
+
+			const result = await this.collection
+				.find({ ownerId: new ObjectId(ownerId), groupId })
+				.toArray();
+			return result.map(
 				(item) =>
 					({
 						id: item._id,
@@ -107,21 +184,36 @@ export default class ExecutionModel extends Model<Execution> {
 						title: item.title
 					} as Execution)
 			);
-        }
-        catch (error) {
-            console.log(error);
-            return [];
-        }
-    }
+		} catch (error) {
+			return {
+				code: 500,
+				message: 'Failed to get all executions by owner ID and group ID'
+			} as Error;
+		}
+	}
 
-    async deleteAllByOwnerId(ownerId: string): Promise<boolean> {
-        try {
-            const result = await this.collection.deleteMany({ ownerId: new ObjectId(ownerId) });
-            return result.deletedCount > 0;
-        }
-        catch (error) {
-            console.log(error);
-            return false;
-        }
-    }
+	async deleteAllByOwnerId(ownerId: string): Promise<boolean | Error> {
+		try {
+			if (!ObjectId.isValid(ownerId)) {
+				return {
+					code: 400,
+					message: 'Invalid Mongo ID'
+				} as Error;
+			}
+
+			const result = await this.collection.deleteMany({ ownerId: new ObjectId(ownerId) });
+			if (result.deletedCount > 0) {
+				return true;
+			}
+			return {
+				code: 404,
+				message: 'Execution not found'
+			} as Error;
+		} catch (error) {
+			return {
+				code: 500,
+				message: 'Failed to delete executions'
+			} as Error;
+		}
+	}
 }
